+++ CHANGELOG.md
@@ -4,13 +4,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## NEXT
+## 0.3.0 - NEXT
 
 ### Changed
 
 * Updated tests
 * Updated to `http-interop/http-middleware#0.4`
 * Updated `friendsofphp/php-cs-fixer#2.0`
+
+### Added
+
+* New method `ContentType::charsets()` to define the available charsets and negotiate the `Accept-Charset` header.
+
+### Fixed
+
+* Only text-based formats (html, text, css, etc) adds the `charset=[charset]` sufix to `Content-Type` header.
 
 ## 0.2.0 - 2016-11-22